WIP: Use color flags for GC.
This is a small step in trying to use a memory bitmap for small GC
objects, rather than using PyGC_Head.  The small step goal is to not
require moving objects to different doubly-linked links during the
collection process (in order to put them in different sets).
Instead, we use color flags and iterate over the whole collection
set.  The GC behavior should be the same after this change although
performance will likely be a little worse.