wip: use color tags for cyclic GC
This changes the GC to use color flags and generation markers rather
than putting objects into different linked-lists.  It is inefficient
at the moment, doing many passes of the whole linked-list.  However,
the long term plan is to use a special version of obmalloc and keep
GC objects in special arenas.  We can use memory bitmaps for the
color and generation flags.  We could use a flag in each arena to
mark if the whole arena contains any objects of interest (e.g. a
per-arena generation marker).  I believe this could be faster and
more memory efficient than the current GC implementation.