wip: gc colors
This removes the use of the double-linked lists for GC (except for
finding all GC objects at start of collection).  The lists could be
replaced by having a heap traversal process that can find all GC
objects.  E.g. bitmaps for the small object arenas and a collection for
the "large" objects.