wip: gc mark alive pass
When doing a full GC pass, first a "mark alive" pass based on known
object roots.  All the objects found in this pass can be excluded from
the cyclic GC work.

A further elaboration would be to do this "mark alive" pass
incrementally, to avoid long GC pauses.  Once the pass reaches the end
of the heap, the full GC cyclic collection could be done on the objects
that are not marked as alive.   It doesn't matter if the "mark alive"
pass is not completely accurate since it only means we defer collecting
some garbage, potentially.

This prototype uses the C stack to traverse the pointers with
tp_traverse.  We could instead use the same linked-list trick the
regular GC pass uses (move things to a different list as we find them
alive and traverse them).